issue #2 remove duplicated code
